Norminette Updates

+++ libft/ft_striteri.c
@@ -6,7 +6,7 @@
 /*   By: abeh <abeh@student.42singapore.sg>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/16 04:24:34 by abeh              #+#    #+#             */
-/*   Updated: 2024/05/27 16:33:05 by abeh             ###   ########.fr       */
+/*   Updated: 2024/05/28 03:44:38 by abeh             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	ft_striteri(char *s, void (*f)(unsigned int, char *))
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i])
